+++ lib/net.dart
@@ -5,5 +5,7 @@ library net;
 
 export 'src/net/email.dart';
 export 'src/net/ipv4.dart';
+export 'src/net/ipv6.dart';
 export 'src/net/req_email.dart';
 export 'src/net/req_ipv4.dart';
+export 'src/net/req_ipv6.dart';